+++ src/finlang.test/TranspilerTest/C99TranspilerTest.cs
@@ -9,7 +9,7 @@ public class C99TranspilerTest
 {
     string slnPath = ExSln2Fixture.GetSlnPath();
     string destDirPath = ExSln2Fixture.GetSlnDir() + "/c99/gen/";
-    Transpiler.Transpiler transpiler;
+    Transpiler.CTranspiler transpiler;
 
     public C99TranspilerTest()
     {